UI, Pipette and Glove Box Prototypes
Added in some basic UI panels introducing the user to the experience. Features poke interactable buttons and a panel screen which follows the user's head.

Additionally, added a box of latex gloves which when pressed, will at a blue material to the user's hands simulating wearing a pair of latex gloves

Finally, added an animation to the pipette plunger, moving down smoothly when it is being held and the trigger is pressed.