Room Validation, Improved Workstation, Health & Safety Agreement
Added a scene which activates on startup which prompts the user to agree to a health and safety agreement, before loading their chosen experiment.

Added room validation which checks if there is a table in their room setup, and if it is of an adequate size.

Added more objects to the scene workstation including a centrifuge, improved blood tubes, and improved hand grab interactions.

Finally, added some basic instruction panels which float above the workstation.