Centrifuge, Instructions and UI, Plasma, Sounds
Centrifuge is nearly finished except for a panel to show countdown timer. Added shaders, a button, 6 slots for blood tubes, an animation which makes it spin, and new blood tubes with the RBCs/Plasma split.

Added bit instructions panels with a manager class which changes the border colour to green and plays a sound effect when an instruction is completed successfully. Operates on a series of boolean values to guide the user through the experience.

Started the instructions for the pipetting section but main panels still to be done.  Still need to add an animation to the plasma and add instructions events.

Added a few sound effects for positive reinforcement stuff