Refactor image and fix microscope
This branch was intended to refactor the game to use Image instead of RawImage. I got distracted and after refactoring, continued to update the microtome activity. The microtome activity now implements a IllnessGallery class, which houses a list of illnesses each with their own sprite, name, treatments and symptoms. These Illnesses are used to populate the content on the laptop screen.

Now, when a slice is spawned, it copies the sprite attached to the main sample block to itself. This means that whenever a new slice is made from the same block the image is always the same.

Added a particule system for the fridge to create some cool steamy visuals and added a bunch of sound effects for various interactions