Add new sample block models. Add challenge mode
Meta quest link completely died on me so progress is now slow. Client has asked for a switch from fridge to a cupboard and some sort of water bath thing. Explained the complexities of doing so and don't know if it is going to be possible. Generally working well and looks great in built version