Add dynamic spawn locations for ffpe block cabinet and wall art
Replaced the fridge with a cabinet as requested by sponsor. The cabinet now spawns wherever the user has marked a STORAGE anchor in the room setup. Wall art panel now spawns on the walls whereever wall art has been marked and displays a tooltip for opening the menu and reloading the scene.