Improve Various UX things
made various small changed to the user experience including better grab interactions with the test tubes, tray of FFPE samples, and sound effects for the pipette and challenge modes. Fixed the lighting issue in blood separator and disabled the centrifuge lid grab interaction whilst it is spinning. Fixed a couple of room validation issues and did some minor refactoring of the centrifuge code to be a bit more readable.